Create Get Location
We complicate the work of our robot! Now he knows how to convert movement commands into the correct signal and move according to it:

"forward" means y + 1 (a step forward);
"back" means y - 1 (step back);
"right" means x + 1 (step to the right);
"left" means x - 1 (step to the left).
But it would be great if the robot knew where it is even without GPS. To do this, implement the get_location() function, which:

Accepts 2 parameters:
list of initial coordinates coordinates in the form [x, y];
a list of commands in the form ["command1", "command2", "command3" ...].
Returns the list of final coordinates [x, y] after movements according to the commands from the commands list.